API-code/requirements.txt
wheel
fastapi
uvicorn
aiofiles
Jinja2
pytube3